requirements.txt
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
PyJWT==2.4.0
python-telegram-bot==13.11
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
gunicorn
Werkzeug
yt-dlp